Split responsibilities between the consumer and the rest of the code
