+++ scala/src/main/scala/com/github/pedrovgs/kafkaplayground/flash/ElasticsearchConsumer.scala
@@ -1,17 +1,10 @@
 package com.github.pedrovgs.kafkaplayground.flash
 
 import cakesolutions.kafka.KafkaConsumer.Conf
-import io.searchbox.client.config.HttpClientConfig
-import io.searchbox.client.{JestClient, JestClientFactory}
-import io.searchbox.core.Index
-import org.apache.http.auth.{AuthScope, UsernamePasswordCredentials}
-import org.apache.http.impl.client.BasicCredentialsProvider
-import org.apache.kafka.clients.consumer.KafkaConsumer
+import com.github.pedrovgs.kafkaplayground.flash.elasticsearch.ElasticClient
 import org.apache.kafka.common.serialization.StringDeserializer
 
-import scala.annotation.tailrec
 import scala.collection.JavaConverters._
-import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
 
 object ElasticsearchConsumer {
@@ -20,11 +13,7 @@ object ElasticsearchConsumer {
 
 class ElasticsearchConsumer(private val brokerAddress: String,
                             private val topic: String,
-                            private val elasticsearchHost: String,
-                            private val elasticsearchUser: String,
-                            private val elasticsearchPass: String,
-                            private val elasticIndex: String,
-                            implicit val ec: ExecutionContext = ExecutionContext.global) {
+                            private val elasticClient: ElasticClient) {
 
   import ElasticsearchConsumer._
 
@@ -33,47 +22,22 @@ class ElasticsearchConsumer(private val brokerAddress: String,
       bootstrapServers = brokerAddress,
       keyDeserializer = new StringDeserializer(),
       valueDeserializer = new StringDeserializer(),
-      groupId = groupId
+      groupId = s"$topic-$groupId"
     )
   )
 
-  private val client: JestClient = {
-    val factory  = new JestClientFactory()
-    val provider = new BasicCredentialsProvider()
-    provider.setCredentials(AuthScope.ANY,
-                            new UsernamePasswordCredentials(elasticsearchUser, elasticsearchPass))
-    factory.setHttpClientConfig(
-      new HttpClientConfig.Builder(elasticsearchHost)
-        .credentialsProvider(provider)
-        .build())
-    factory.getObject
-  }
+  consumer.subscribe(List(topic).asJava)
 
-  def start(): Unit = {
-    consumer.subscribe(List(topic).asJava)
-    sendTweetsInfoToElasticsearch(consumer)
-  }
-
-  @tailrec
-  private def sendTweetsInfoToElasticsearch(consumer: KafkaConsumer[String, String]): Unit = {
-    val records = consumer.poll(10.seconds.toMillis)
+  def poll(): Unit = {
+    println(s"Polling messages from the kafka consumer at topic: $topic.")
+    val records = consumer.poll(5.seconds.toMillis)
+    println(s"We've fetched ${records.count()} records.")
     records.forEach { record =>
-      val id = s"${record.topic()}_${record.partition()}_${record.offset()}"
+      val id      = s"${record.topic()}_${record.partition()}_${record.offset()}"
       val content = record.value()
-      println(s"Saving topic content into elastic: $content")
-      saveContentIntoElasticsearch(id, content)
+      println(s"Saving content from the topic $topic content into elastic: $content")
+      elasticClient.insertOrUpdate(id, content)
     }
-    sendTweetsInfoToElasticsearch(consumer)
   }
 
-  private def saveContentIntoElasticsearch(id: String, content: String) = {
-    try {
-      val index  = new Index.Builder(content).index(elasticIndex).`type`("tweets").build
-      val result = client.execute(index)
-      println(s"Elasticsearch new document id: ${result.getId}")
-    } catch {
-      case e: Exception =>
-        println(e.getMessage)
-    }
-  }
 }
